requirements.txt
PyYAML>=6.0.1
ccxt>=4.3.0
numpy>=1.26.4
pandas>=2.2.2
plotext>=5.3.2
python-dotenv>=1.0.1
requests>=2.31.0
rich>=13.7.1
websockets>=12.0
